+++ src/DateTime.cpp
@@ -11,7 +11,7 @@ Time time_now()
 
 double time_elapsed_ms(Time current, Time then)
 {
-  double ms = std::chrono::duration_cast<std::chrono::milliseconds>(current - then).count();
+  double ms = (double)std::chrono::duration_cast<std::chrono::milliseconds>(current - then).count();
   return ms;
 }
